pyproject.toml
[project]
name = "restaurant-agent"
version = "0.1.0"
description = "Multi-agent restaurant management system using Google ADK with MCP"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "google-adk>=1.19.0",
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
    "fastmcp>=2.13.1",
    "absl-py>=2.2.1",
    "google-cloud-aiplatform[agent_engines]>=1.91.0",
    "httpx>=0.28.0",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["restaurant_agent"]

